+++ docs/changelog.md
@@ -177,3 +177,9 @@
 - No template available
 	- Print a message when the lenght template list is 0
 
+**Enhancement**
+
+- Local install template
+	- Add a local installation of templates
+	- Add an update option to upgrade local templates from local source
+